Fixing up code
Adding comments.
Creating parent to Child and to Teacher page by adding its own different page.
Created stylesheets to make it look somewhat nice but not the final look jst for now,, :P
Forgot to mention fixed design and layout for teach and Parent sign in~ Hopefully this commits well..